Move the namespace-neutral error vocabulary into its own leaf
memoryops needed one sentinel — ErrValidation — and the value the whole repo
matches lived in issueops. So the memory leaf imported the issue package for
one errors.New and pulled internal/types in behind it. That import made a
claim that is false: memory is not downstream of issues, the two are sibling
planes over one config table.

beadserrors now declares the vocabulary that belongs to no plane in
particular: ErrValidation, ErrNotFound, ErrNotInitialized and the
ErrUnsupported type. issueops re-exports them by alias exactly as
internal/storage and backend already re-export from issueops, so every value
is identical and nothing moved: 141 qualified ErrValidation references across
the tree, zero edits. memoryops' whole beads-internal dependency set is now
beadserrors, where it was issueops + internal/types.

The test for what belongs down there is whether a leaf for a plane nobody has
written yet would need it. A request can be invalid, a row can be missing, a
substrate can be uninitialized and a backend can not implement a capability,
on any plane. ErrCloseBlocked and ErrDependencyCycle name issue concepts and
stay in issueops.

beadserrors imports stdlib and nothing else, enforced by a strict depguard
rule rather than asked for in a comment: it sits beneath every leaf, so
whatever it imports is imported by all of them, and the next tempting import
is always one shared constant that "obviously belongs with the errors".

ADDING_AN_ISSUEOPS_ROLE.md carried a wrong cost model for this move — that
doing it later meant "every errors.Is site moves at once". A Go alias
preserves identity in both directions, so the new leaf declares canonically
and the old home aliases forward. What actually grows per namespace is the
number of leaves that have baked in the wrong dependency direction, not
migration churn. Corrected in place rather than deleted, since the reasoning
is what a third namespace inherits.
